modify_anu

+++ application/controllers/Leads.php
@@ -13,16 +13,45 @@ class Leads extends CI_Controller
         $this->output->set_header('Cache-Control: no-store, no-cache, must-revalidate, post-check=0, pre-check=0');
         $this->output->set_header('Pragma: no-cache');
 
-        $this->user_model->check_session_data('admin');
+        
     }
 
+public function leads_data()
+ {
+  $data['leads_data'] = $this->leads_model->export_data();
+  //$this->load->view('backend/admin/lead/lead_view', $data);
+ }
 
+ function export()
+ {
+    if ($this->session->userdata('is_sales_manager') == true || $this->session->userdata('is_sales_team') == true || $this->session->userdata('is_campaign_team') == true ) {
+
+        $this->session->set_flashdata('error_message', get_phrase('restricted'));
+        redirect(site_url('leads/leads/leads_view'), 'refresh');
+    }
+    $file_name = 'leads_details_on_'.date('Ymd').'.csv'; 
+     header("Content-Description: File Transfer"); 
+     header("Content-Disposition: attachment; filename=$file_name"); 
+     header("Content-Type: application/csv;");
+   
+     // get data 
+     $leads_data = $this->leads_model->export_data();
+
+     // file creation 
+     $file = fopen('php://output', 'w');
+     $header = array("name, email, website, phonenumber, designation, company, country_id, city_id, category_id, course_id"); 
+     fputcsv($file, $header);
+     foreach ($leads_data->result_array() as $key => $value)
+     { 
+       fputcsv($file, $value); 
+     }
+     fclose($file); 
+     exit; 
+ }
 
     public function leadssource($param1 = "", $param2 = "")
     {
-        if ($this->session->userdata('admin_login') != true) {
-            redirect(site_url('login'), 'refresh');
-        }
+      
         
         // CHECK ACCESS PERMISSION
         check_permission('leads');
@@ -35,7 +64,7 @@ class Leads extends CI_Controller
             } else {
                 $this->session->set_flashdata('error_message', get_phrase('Source_name_already_exists'));
             }
-            redirect(site_url('leads/leadssource/add_leadssource'), 'refresh');
+            redirect(site_url('leads/leadssource/leadssource'), 'refresh');
         } 
 
 
@@ -47,15 +76,21 @@ class Leads extends CI_Controller
             } else {
                 $this->session->set_flashdata('error_message', get_phrase('city_name_already_exists'));
             }
-            redirect(site_url('leads/leadssource/add_leadssource'), 'refresh');
+            redirect(site_url('leads/leadssource/leadssource'), 'refresh');
 
             
 
         } elseif ($param1 == "delete") {
 
-            $this->leads_model->delete_leads_source($param2);
-            $this->session->set_flashdata('flash_message', get_phrase('data_deleted'));
-            redirect(site_url('leads/leadssource/add_leadssource'), 'refresh');
+            if ($this->session->userdata('admin_login') == true) {
+
+                $this->leads_model->delete_leads_source($param2);
+                $this->session->set_flashdata('flash_message', get_phrase('data_deleted'));
+                redirect(site_url('leads/leadssource/leadssource'), 'refresh');
+            }
+            $this->session->set_flashdata('error_message', get_phrase('restricted'));
+            redirect(site_url('leads/leadssource/leadssource'), 'refresh');
+           
         }
 
         
@@ -69,9 +104,7 @@ class Leads extends CI_Controller
 
     public function leadssource_form($param1 = "", $param2 = "")
     {
-        if ($this->session->userdata('admin_login') != true) {
-            redirect(site_url('login'), 'refresh');
-        }
+      
 
         // CHECK ACCESS PERMISSION
         check_permission('leads');
@@ -104,9 +137,7 @@ class Leads extends CI_Controller
 
     public function leadsstatus($param1 = "", $param2 = "")    {
 
-        if ($this->session->userdata('admin_login') != true) {
-            redirect(site_url('login'), 'refresh');
-        }
+    
 
 
          // CHECK ACCESS PERMISSION
@@ -121,7 +152,7 @@ class Leads extends CI_Controller
             } else {
                 $this->session->set_flashdata('error_message', get_phrase('Source_status_already_exists'));
             }
-            redirect(site_url('leads/leadsstatus/add_leadsstatus'), 'refresh');
+            redirect(site_url('leads/leadsstatus/leadsstatus'), 'refresh');
         } 
 
         elseif ($param1 == "edit") {
@@ -132,7 +163,7 @@ class Leads extends CI_Controller
             } else {
                 $this->session->set_flashdata('error_message', get_phrase('Source_status_already_exists'));
             }
-            redirect(site_url('leads/leadsstatus/add_leadsstatus'), 'refresh');
+            redirect(site_url('leads/leadsstatus/leadsstatus'), 'refresh');
 
             
 
@@ -143,7 +174,7 @@ class Leads extends CI_Controller
 
             $this->leads_model->delete_leads_status($param2);
             $this->session->set_flashdata('flash_message', get_phrase('data_deleted'));
-            redirect(site_url('leads/leadsstatus/add_leadsstatus'), 'refresh');
+            redirect(site_url('leads/leadsstatus/leadsstatus'), 'refresh');
         }
 
 
@@ -158,9 +189,6 @@ class Leads extends CI_Controller
 
     public function leadsstatus_form($param1 = "", $param2 = "")
     {
-        if ($this->session->userdata('admin_login') != true) {
-            redirect(site_url('login'), 'refresh');
-        }
 
         // CHECK ACCESS PERMISSION
         check_permission('leads');
@@ -190,23 +218,26 @@ class Leads extends CI_Controller
 
     public function leads($param1 = "", $param2 = "")    {
 
-        if ($this->session->userdata('admin_login') != true) {
-            redirect(site_url('login'), 'refresh');
-        }
+       
 
   // CHECK ACCESS PERMISSION
   check_permission('leads');
 
 
+
   if ($param1 == 'add') {
 
+
     $response = $this->leads_model->addleads();
+
+    
     if ($response) {
+        
         $this->session->set_flashdata('flash_message', get_phrase('data_added_successfully'));
     } else {
         $this->session->set_flashdata('error_message', get_phrase('leads_status_already_exists'));
     }
-    redirect(site_url('leads/leads/leads'), 'refresh');
+    redirect(site_url('leads/leads/leads_view'), 'refresh');
 } 
 elseif ($param1 == "edit") {
 
@@ -216,13 +247,20 @@ elseif ($param1 == "edit") {
     } else {
         $this->session->set_flashdata('error_message', get_phrase('leads_name_already_exists'));
     }
-    redirect(site_url('leads/leads/leads'), 'refresh');
+    redirect(site_url('leads/leads/leads_view'), 'refresh');
 }
  elseif ($param1 == "delete") {
 
+    if ($this->session->userdata('admin_login') == true) {
+
     $this->leads_model->delete_leads($param2);
     $this->session->set_flashdata('flash_message', get_phrase('data_deleted'));
-    redirect(site_url('leads/leads/leads'), 'refresh');
+    redirect(site_url('leads/leads/leads_view'), 'refresh');
+
+    
+ }
+ $this->session->set_flashdata('error_message', get_phrase('restricted'));
+ redirect(site_url('leads/leads/leads_view'), 'refresh');
 }
 
   
@@ -236,21 +274,35 @@ elseif ($param1 == "edit") {
 
 public function leads_form($param1 = "", $param2 = "")
 {
-    if ($this->session->userdata('admin_login') != true) {
-        redirect(site_url('login'), 'refresh');
-    }
+    
 
     // CHECK ACCESS PERMISSION
     check_permission('leads');
 
+ 
+
     if ($param1 == "addleads") {
+
+        if ($this->session->userdata('is_sales_manager') == true || $this->session->userdata('is_sales_team') == true ) {
+
+            $this->session->set_flashdata('error_message', get_phrase('restricted'));
+            redirect(site_url('leads/leads/leads_view'), 'refresh');
+        }
 
         $page_data['page_name'] = 'lead/leads';
         $page_data['leads'] = $this->leads_model->get_leads()->result_array();
         $page_data['page_title'] = get_phrase('leads');
+      
+       
     }
 
     if ($param1 == "edit_leads") {
+
+        if ($this->session->userdata('is_campaign_team') == true || $this->session->userdata('is_sales_team') == true ) {
+
+            $this->session->set_flashdata('error_message', get_phrase('restricted'));
+            redirect(site_url('leads/leads/leads_view'), 'refresh');
+        }
 
         $page_data['page_name'] = 'lead/leads_edit';
         $page_data['page_title'] = get_phrase('edit_leads');
